requirements.txt
# LUAD Precision Oncology Platform — Python dependencies
# For Streamlit dashboard (streamlit_app.py)
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0

# For pipeline modules (01-07)
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
requests>=2.28.0
biopython>=1.81
gseapy>=1.0.0
lifelines>=0.27.0

# For module 06 ESM2 (optional, GPU recommended)
# torch>=2.0.0
# transformers>=4.35.0
# fair-esm>=2.0.0
